requirements.txt
numpy==1.18.1
parasail==1.1.19
dnaio==0.4.1
editdistance==0.5.3
openpyxl==2.4.8
requests==2.23.0
lxml>=4.1.1
plotly==4.5.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-cloud-logging==1.15.0
boto3==1.12.16
Astrology-Aries
